Add API handling for ATAT eligibility endpoint
Depends on #8606

This patch adds functionality into Calypso to fetch and handle requests
to the automated-transfer eligibility endpoint on the WordPress.com API.

The functionality can be tested by issuing a simple dispatch from the
browser console with the given site domain.

```js
dispatch( {
	type: actionTypes.AUTOMATED_TRANSFER_ELIGIBILITY_REQUEST,
	domain: 'wordbaykun.com'
} );
```

An API requests should execute and return the elibility information in a
normalized fashion.

Note that for now I have elected to make a simple `fetch` request
instead of building this functionality into `wpcom.js`. This was done
because we can contain all of the API-specific code in this new
middleware directory and we don't need it in `wpcom.js`. This pattern
also feeds directly into an HTTP middleware which could issue on the URL
and associated data instead of directly making the requests here.